+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 
 
+## Unreleased
+
+### Bug Fix
+- Fix tabbar left right are reversed in Firefox 29 (Tested with Classic Theme Restorer 1.1.9)
+- Fix vertical tabbar splitter not showing on Firefox 29 on startup
+
+### Known Bugs
+- Due to the bug fix above, the tab bar position is reversed on Firefox 28 or below (Won't fix)
+
+
 ## 0.11.1
 
 ### Bug Fix